feat: 여행 일자 삭제 후 재정렬
여행 일자 삭제에 아래 항목 추가
1. 여행 일자 삭제 실행 전 여행 일자의 indexSort 조회
2. 여행 스케줄 및 일자 삭제 완료 후 삭제된 여행 일자보다 indexSort가 큰 데이터들의 값 재정렬

Fixes #DEV-70

+++ src/main/java/com/example/plana/service/TripService.java
@@ -307,10 +307,13 @@ public class TripService {
         Map<String, Object> deleteParams = new HashMap<>();
         deleteParams.put("tripId", tripId);
         deleteParams.put("tripDayId", tripDayId);
+        // 1. SELECT TRIP DAY - getIndexSort
+        int dayIndexSort = 0;
+        dayIndexSort = tripMapper.readTripDayIndexSort(tripDayId);
+        if (dayIndexSort <= 0) throw new BusinessException(ErrorCode.TRIP_DAY_NOT_FOUND);
 
+        // 2. 여행 스케줄 전체 삭제
         int result = -1;
-
-        // 1. 여행 스케줄 전체 삭제
         try {
             result = tripMapper.deleteTripSchedulesByTripDayId(deleteParams);
         } catch (Exception e) {
@@ -320,13 +323,26 @@ public class TripService {
             throw new BusinessException(ErrorCode.TRIP_SCHEDULE_NOT_FOUND);
         }
 
+        // 3. 여행 일자 삭제
         result = -1;
-
-        // 2. 여행 일자 삭제
         try {
             result = tripMapper.deleteTripDay(deleteParams);
         } catch (Exception e) {
             throw new BusinessException(ErrorCode.TRIP_DAY_DELETE_FAILED);
+        }
+        if (result == 0) {
+            throw new BusinessException(ErrorCode.TRIP_DAY_NOT_FOUND);
+        }
+
+        // 4. 삭제된 일자 이후의 index sort 재정렬
+        result = -1;
+        Map<String, Object> reorderParams = new HashMap<>();
+        reorderParams.put("tripId", tripId);
+        reorderParams.put("indexSort", dayIndexSort);
+        try {
+            result = tripMapper.updateTripDaysIndexSortAfterDelete(reorderParams);
+        } catch (Exception e) {
+            throw new BusinessException(ErrorCode.TRIP_DAY_REORDER_FAILED);
         }
         if (result == 0) {
             throw new BusinessException(ErrorCode.TRIP_DAY_NOT_FOUND);
